requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
argilla==1.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
cohere==4.3.0
commonmark==0.9.1
cryptography==40.0.2
dataclasses-json==0.5.7
decorator==5.1.1
Deprecated==1.2.13
entrypoints==0.4
et-xmlfile==1.1.0
faiss-cpu==1.7.4
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
jsonlines==3.1.0
jsonschema==4.17.3
langchain==0.0.142
llama-cpp-python==0.1.41
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
nomic==1.1.6
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
protobuf==3.20.3
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
pygpt4all==1.0.1
pygptj==1.0.10
pyllamacpp==1.0.6
Pympler==1.0.1
pypandoc==1.11
pypdf==3.8.1
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
rfc3986==1.5.0
rich==13.0.1
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.12
streamlit==1.22.0
streamlit-ace==0.1.1
sympy==1.11.1
tenacity==8.2.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
tornado==6.3.1
tqdm==4.65.0
transformers==4.28.1
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
unstructured==0.6.5
urllib3==1.26.15
validators==0.20.0
wonderwords==2.2.0
wrapt==1.14.1
XlsxWriter==3.1.0
yarl==1.9.2
zipp==3.15.0
farm-haystack
torchaudio
sentence-transformers
